Move blebox base entity to separate module
